[misc] sync with changes of logger in ppl.common

+++ src/ppl/nn/engines/arm/runtime_arm_device.cc
@@ -92,7 +92,7 @@ void RuntimeArmDevice::FreeTmpBuffer(BufferDesc* buffer) {
 
 RetCode RuntimeArmDevice::Configure(uint32_t option, ...) {
     if (option >= DEV_CONF_MAX) {
-        LOG(ERROR) << "invalid option[" << option << "] >= [" << DEV_CONF_MAX << "]";
+        LOG(ERROR) << "invalid option[" << option << "] >= [" << (uint32_t)DEV_CONF_MAX << "]";
         return RC_INVALID_VALUE;
     }
