+++ src/ppl/nn/engines/cuda/engine.cc
@@ -360,7 +360,7 @@ CudaEngine::ConfHandlerFunc CudaEngine::conf_handlers_[] = {
 
 RetCode CudaEngine::Configure(uint32_t option, ...) {
     if (option >= ENGINE_CONF_MAX) {
-        LOG(ERROR) << "invalid option[" << option << "] >= [" << ENGINE_CONF_MAX << "]";
+        LOG(ERROR) << "invalid option[" << option << "] >= [" << (uint32_t)ENGINE_CONF_MAX << "]";
         return RC_INVALID_VALUE;
     }
