+++ src/ppl/nn/engines/riscv/runtime_riscv_device.cc
@@ -93,7 +93,7 @@ void RuntimeRiscvDevice::FreeTmpBuffer(BufferDesc* buffer) {
 
 RetCode RuntimeRiscvDevice::Configure(uint32_t option, ...) {
     if (option >= DEV_CONF_MAX) {
-        LOG(ERROR) << "invalid option[" << option << "] >= [" << DEV_CONF_MAX << "]";
+        LOG(ERROR) << "invalid option[" << option << "] >= [" << (uint32_t)DEV_CONF_MAX << "]";
         return RC_INVALID_VALUE;
     }
